zenaida1/requirements/requirements-testing.txt
tox
isort
flake8
mock
django-debug-toolbar
pkgversion

# PyTest for running the tests.
pytest
pytest-cov
pytest-django

# docs
sphinx
# sphinxcontrib-plantuml
sphinx_rtd_theme
